# docs/swag/wallet/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the Wallet API server",
        "title": "Wallet API Documentation",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "ap11.0dev.cc",
    "basePath": "/",
    "paths": {
...
        "/wallet/create": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new wallet for the authenticated user if they don't have one",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Create new wallet",
                "responses": {
                    "200": {
                        "description": "Successfully created wallet",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Wallet already exists",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
        "/wallet/detail": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve detailed information about the authenticated user's wallet including balance and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet details",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved wallet details",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - User ID not found in context",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found - Wallet not found for user",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/apiresp.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "token",
            "in": "header"
        }
    }
}